finsight/requirements.txt
streamlit
pandas
numpy
plotly
yfinance
scipy
fpdf2
kaleido
